Add x5c header to response tokens
